Change backpropagation
lr is change from 1e-10 to 1e-9 to speed up test pass.
Remove use of DFS. The compute graph is already in memory based on the backprapagation's root node and parsable using LeftChildren / RightChildren as path conbination.
Back propagation now begin from root of backpropagation and follow the graph computation in memory. Parsing is done from LefChildren to RightChildren

root of backpropagation is the Value where Backpropagate is call. Here is loos.Backpropagate()